+++ tests/e2e/game/index.js
@@ -12,4 +12,5 @@ describe("E2E - 🎲 Game tests", () => {
     require("./no-winner-game.test");
     require("./game-won-by-lovers.test");
     require("./game-options.test");
+    require("./ancient-revenge-game.test");
 });