+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#73](https://github.com/antoinezanardi/werewolves-assistant-api/issues/73) - Vile Father of Wolves Role implemented.
 * [#86](https://github.com/antoinezanardi/werewolves-assistant-api/issues/86) - Extend player limit to 40.
 * [#87](https://github.com/antoinezanardi/werewolves-assistant-api/issues/87) - Add isRevealed to role structure.
+* [#88](https://github.com/antoinezanardi/werewolves-assistant-api/issues/88) - Ancient Role implemented.
 * [#91](https://github.com/antoinezanardi/werewolves-assistant-api/issues/91) - Game option for no sheriff.
 * [#95](https://github.com/antoinezanardi/werewolves-assistant-api/issues/95) - Limit game history length.
 * [#96](https://github.com/antoinezanardi/werewolves-assistant-api/issues/96) - Add dead players in game history entries.
@@ -30,11 +31,13 @@
 ### 📦 Packages
 
 * `deepmerge` installed with version `4.2.2`.
+* `@sentry/node` updated to version `5.30.0`.
 * `apidoc` updated to version `0.26.0`.
 * `eslint` updated to version `7.17.0`.
 * `express-validator` updated to version `6.9.2`.
 * `mongoose` updated to version `5.11.11`.
 * `nodemon` updated to version `2.0.7`.
+* `qs` updated to version `6.9.5`.
 
 ---